reorganized the discord command folder structure
