+++ src/main/java/com/ambrosia/add/discord/DiscordModule.java
@@ -4,15 +4,15 @@ import apple.lib.modules.AppleModule;
 import apple.lib.modules.configs.data.config.AppleConfig.Builder;
 import apple.lib.modules.configs.factory.AppleConfigLike;
 import com.ambrosia.add.discord.commands.CommandHelp;
-import com.ambrosia.add.discord.commands.casino.CommandCasino;
-import com.ambrosia.add.discord.commands.delete.CommandDelete;
-import com.ambrosia.add.discord.commands.operation.CommandCash;
-import com.ambrosia.add.discord.commands.profile.CommandLink;
-import com.ambrosia.add.discord.commands.profile.CreateProfileCommand;
-import com.ambrosia.add.discord.commands.profile.ProfileCommand;
-import com.ambrosia.add.discord.commands.profile.ViewProfileCommand;
-import com.ambrosia.add.discord.commands.restart.CommandRestart;
-import com.ambrosia.add.discord.commands.stats.CommandStats;
+import com.ambrosia.add.discord.commands.manager.casino.CommandCasino;
+import com.ambrosia.add.discord.commands.manager.delete.CommandDelete;
+import com.ambrosia.add.discord.commands.dealer.cash.CommandCash;
+import com.ambrosia.add.discord.commands.player.profile.CommandLink;
+import com.ambrosia.add.discord.commands.player.profile.CreateProfileCommand;
+import com.ambrosia.add.discord.commands.player.profile.ProfileCommand;
+import com.ambrosia.add.discord.commands.player.profile.ViewProfileCommand;
+import com.ambrosia.add.discord.commands.manager.restart.CommandRestart;
+import com.ambrosia.add.discord.commands.player.stats.CommandStats;
 import com.ambrosia.add.discord.log.DiscordLog;
 import com.ambrosia.add.discord.log.RestartingMessageManager;
 import com.github.AndrewAlbizati.Blackjack;
