--- src/main/java/com/ambrosia/add/discord/commands/casino/ratio/GuiCasinoRatio.java
@@ -1,18 +1,0 @@
-package com.ambrosia.add.discord.commands.casino.ratio;
-
-import discord.util.dcf.gui.base.gui.DCFGui;
-import discord.util.dcf.gui.base.page.DCFGuiPage;
-import net.dv8tion.jda.api.utils.messages.MessageCreateData;
-
-public class GuiCasinoRatio extends DCFGuiPage<DCFGui> {
-
-    public GuiCasinoRatio(DCFGui dcfGui) {
-        super(dcfGui);
-    }
-
-    @Override
-    public MessageCreateData makeMessage() {
-        return null;
-    }
-
-}
